Back/Python_extensions/requirement.txt
faiss-cpu
numpy
Pillow
cmake
dlib
face-recognition
transformers
onnxruntime


flask